+++ inspect/inspectors/port_scanner.go
@@ -36,7 +36,7 @@ func (p *PortScanner) Name() string {
 }
 
 func (p *PortScanner) Inspect(g *graph.Graph) error {
-	sgroups, err := g.GetAllResources(cloud.SecGroup)
+	sgroups, err := g.GetAllResources(cloud.SecurityGroup)
 	if err != nil {
 		return err
 	}
@@ -72,7 +72,7 @@ func (p *PortScanner) Print(w io.Writer) {
 		} else {
 			targets = strings.Join(p.applyingOn[sg], ", ")
 		}
-		fmt.Fprintf(w, "Secgroup %s applying on %s: \n", sg, targets)
+		fmt.Fprintf(w, "Securitygroup %s applying on %s: \n", sg, targets)
 
 		var allPermissive bool